add 'create a tutorial' button to TutorialEndCard, refactor iconStyle variable
